182403149 - Aggregate Student Data cleanup
- move & rename files.
- remove `console.log` debugging.
- remove unused comments.

[#182403149]
https://www.pivotaltracker.com/story/show/182403149
